backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/GetClustersByStoragePoolIdQuery.java
package org.ovirt.engine.core.bll;

import org.ovirt.engine.core.common.queries.IdQueryParameters;

public class GetClustersByStoragePoolIdQuery<P extends IdQueryParameters>
        extends QueriesCommandBase<P> {
    public GetClustersByStoragePoolIdQuery(P parameters) {
        super(parameters);
    }

    @Override
    protected void executeQueryCommand() {
        getQueryReturnValue().setReturnValue(
                getDbFacade().getClusterDao().getAllForStoragePool(
                        getParameters().getId(),
                        getUserID(),
                        getParameters().isFiltered()));
    }
}
